parser bug fix + norminette resolved fullt

+++ libft/base_nuber.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   base_nuber.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: apatvaka <apatvaka@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rbarkhud <rbarkhud@student.42yerevan.am    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/03 14:15:31 by alen              #+#    #+#             */
-/*   Updated: 2025/04/06 17:15:30 by apatvaka         ###   ########.fr       */
+/*   Updated: 2026/01/25 18:59:03 by rbarkhud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	base_number(char *base, unsigned long num)
 	int	count;
 
 	count = 0;
-	if (num >= ft_strlen(base))
+	if (num >= (size_t)ft_strlen(base))
 		count += base_number(base, num / ft_strlen(base));
 	ft_putchar(base[num % ft_strlen(base)]);
 	count++;
