+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: rbarkhud <rbarkhud@student.42yerevan.am    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/14 15:13:07 by apatvaka          #+#    #+#             */
-/*   Updated: 2025/08/27 20:08:16 by rbarkhud         ###   ########.fr       */
+/*   Updated: 2026/01/25 18:56:52 by rbarkhud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	const char	*c;
